+++ packages/terra-core-docs/src/terra-dev-site/test/tag/HrefTag.test.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import Tag from 'terra-tag';
+
+const HrefTag = () => <Tag id="href" text="Href Tag" href="www.google.com" />;
+
+export default HrefTag;