+++ packages/terra-core-docs/src/terra-dev-site/test/tag/IconAndTextTag.test.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import IconBookmark from 'terra-icon/lib/icon/IconTile';
+import Tag from 'terra-tag';
+
+const icon = <IconBookmark />;
+
+const IconTag = () => <Tag icon={icon} text="Icon Tag" id="iconTag" />;
+
+export default IconTag;