+++ packages/terra-visually-hidden-text/src/terra-dev-site/test/visually-hidden-text/DefaultVisuallyHiddenText.test.jsx
@@ -1,9 +1,0 @@
-import React from 'react';
-import VisuallyHiddenText from '../../../VisuallyHiddenText';
-
-export default () => (
-  <p>
-    No Screen Reader text is added on this page, and should be blank other than this line.
-    <VisuallyHiddenText />
-  </p>
-);